requirements.txt
pymongo
pandas
numpy
tqdm
matplotlib
python-dateutil
python-dotenv
